通常の射撃の種類に応じて間隔変更・半円形移動の射撃間隔変更・体力によってテーブル終了時の待機時間変更・プレイヤーの弾に生存時間追加

+++ DirectXGame/Application/GameObject/Boss/System/BossSystem.cpp
@@ -10,7 +10,7 @@
 
 uint32_t BossSystemContext::IBullet::sSerialNumber = 0;
 
-void BossSystemContext::HealthManager::TakeDamage(int32_t damage)
+void BossSystemContext::HealthManager::TakeDamage(float damage)
 {
 	currentHealth_ -= damage;
 	// 死亡処理
@@ -20,7 +20,7 @@ void BossSystemContext::HealthManager::TakeDamage(int32_t damage)
 	}
 }
 
-void BossSystemContext::HealthManager::Heal(int32_t heal)
+void BossSystemContext::HealthManager::Heal(float heal)
 {
 	if (currentHealth_ < maxHealth_) {
 		currentHealth_ += heal;
